requirements.txt
wikipedia-api
sentence-transformers
faiss-cpu
Flask